PCI: mvebu: Wait for PCIe link to reset after setting device number
After setting the local device nr, the PCIe link may go down briefly. If we
don't wait for the link to come back up we may miss that device when we scan
the bus (assuming it is not present).

This patch waits up to 40ms for the link to go down, and if it did go down,
waits up to 100ms for it to come back up. If the link was down to begin with,
it will not wait at all.

Signed-off-by: Neil Greatorex <neil@fatboyfat.co.uk>